Revert "Work has been paused indefinitely. Was in the middle of separating client and server"
This reverts commit 2120830e

I decided to finish the job without making it a web app. That was too ambitious. I'd like to first finish the project, and then see if I can rework a ton of stuff to go online.

We don't want feature creep, do we? We already have SOUNDS which we *really* did not need.